calloc_arr stddef

+++ calloc_array/my_calloc.c
@@ -1,19 +1,20 @@
 #include <stdlib.h>
+#include <stddef.h>
 
-int *my_calloc(size_t size, int init) 
+int *my_calloc(size_t size, int init)
 {
-    if (size == 0) 
+    if (size == 0)
     {
         return NULL;
     }
 
     int *call = malloc(sizeof(int) * size);
-    if (!call) 
+    if (!call)
     {
         return NULL;
     }
 
-    for (int i = 0; i < size; i++) 
+    for (size_t i = 0; i < size; i++)
     {
         call[i] = init;
     }
